PoC_IA/requirements.txt
langchain==0.3.0
langchain-openai
langchain-google-genai
langchain-community
pinecone
python-dotenv
# streamlit
pypdf
tiktoken
faiss-cpu
openpyxl
xlsxwriter
langchain-pinecone